DefaultImage is now 16:9, linked not duplicated and referenced in xaml
16:9 is a better fit for almost all engines and carriages.

The Image used to be both in the resource folder and in the TkMobile project, now it's a link from the the project to the resource folder.

The ImageResourceExtension is as described in the offical Xamarin Forms documentation in order to load an image in XAML from resources. The Purpose of this is to avoid setting the default Image after the Page is already open, not changing the layout.